refactor: Reorganize all test files to be co-located with source
- Move all tests from test/ directory to src/
- Move tests from __tests__ subdirectories to be next to source files
- Fix naming inconsistencies:
  - error-handling.test.ts → error-handler.test.ts (matches source)
- Split large test file for better organization:
  - realtime-stats.test.ts → 3 separate test files for each module
- Update jest.config.cjs to only search in src directory
- Fix all import paths in test files
- Update CLAUDE.md with test organization guidelines
- Update CONTRIBUTING.md test paths

Benefits:
- Improved visibility and discoverability of tests
- Reduced cognitive load when working with code and tests
- Consistent naming and organization patterns
- All 101 tests continue to pass
